requirements.txt
# Core database dependencies
sqlmodel>=0.0.14
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
pydantic-settings>=2.1.0

# FastAPI and web server
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
requests>=2.31.0
aiofiles>=23.2.1

# Authentication and security
python-jose[cryptography]>=3.3.0
supabase>=2.3.0

# Database migrations
alembic>=1.13.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0

